src/kec_biomat_api/requirements.txt
beautifulsoup4
black
fastapi
faiss-cpu
google-cloud-aiplatform
google-cloud-bigquery
html2text
langchain
langchain-community
langchain-google-vertexai
lxml
markdown
mypy
networkx
nltk
numpy
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-sdk
pandas
pdf2image
prometheus-fastapi-instrumentator
psutil
pydantic
pydantic-settings
pyjwt
pytest
pytest-asyncio
pytesseract
pyyaml
redis
redoc
requests
ruff
scikit-learn
sentence-transformers
starlette
strawberry-graphql[fastapi]
tiktoken
unstructured
uvicorn
